+++ src/lib.rs
@@ -136,6 +136,7 @@ pub use crate::algorithms::constructors::wheel::{WheelMode, wheel_graph};
 pub use crate::algorithms::flow::edge_connectivity::{adhesion, edge_connectivity};
 pub use crate::algorithms::flow::edge_disjoint_paths::edge_disjoint_paths;
 pub use crate::algorithms::flow::max_flow::max_flow_value;
+pub use crate::algorithms::flow::mincut_value::mincut_value;
 pub use crate::algorithms::flow::st_edge_connectivity::st_edge_connectivity;
 pub use crate::algorithms::flow::st_mincut::st_mincut_value;
 pub use crate::algorithms::flow::st_vertex_connectivity::{VconnNei, st_vertex_connectivity};